requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow
python-dotenv
Werkzeug
waitress
